requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
requests==2.28.1
requests-oauthlib==1.3.1
-e ./threader
tqdm==4.64.1
TwitterAPI==2.7.13
urllib3==1.26.12
